django_blog/requirements.txt
Django==4.2.7
django-taggit==4.0.0
Pillow==10.0.1
